backend/requirements.txt
# django
django==3.1.0
djangorestframework==3.11.1
django-extensions==3.0.5
# drf-yasg==1.17.1

# uwsgi
uwsgi==2.0.18

# celery
celery==4.4.6
django-redis==4.12.1
flower==0.9.7
supervisor==4.2.1
django-celery-beat==2.1.0

#crawl
numpy==1.19.1
pandas==0.24.2
requests==2.22.0
lxml==4.4.1
redis==3.5.3
scraperapi-sdk==0.2.2
beautifulsoup4==4.9.3
html5lib

#model
sqlalchemy==1.3.21
psycopg2==2.8.5

